requirements.txt
numpy==1.14.5
tqdm==4.23.4
tensorboardX==1.6
pytorch-pretrained-bert==0.6.1
GitPython==3.0.5
